rand: resolve cov-scan warning
CID 442766 (#1 of 1): Use of 32-bit time_t (Y2K38_SAFETY)
1. store_truncates_time_t: A time_t value is stored in an
integer with too few bits to accommodate it. The expression
time(NULL) is cast to unsigned int.

+++ src/rand.c
@@ -27,7 +27,7 @@
 
 static unsigned int HXrand_obtain_seed(void)
 {
-	unsigned int s;
+	unsigned long s;
 
 #if defined(HAVE_CLOCK_GETTIME)
 	struct timespec tv;
